Improve admin
Often times on sites with large numbers of users
the admin can timeout. Therefore, the user field
was made into a raw_id_field. In addition finding
all logs for a user can be very useful and if
the site uses email addresses as the primary
identifier, then being able to search by email
address is very helpful.